+++ www/Library/admin_widgets/Destroy/res/boolive.Destroy/boolive.Destroy.js
@@ -22,7 +22,7 @@
 
             self.element.find('.submit').click(function(e){
                 e.preventDefault();
-                self.callServer('delete', {object: self.options.object}, function(result, textStatus, jqXHR){
+                self.callServer('destroy', {object: self.options.object}, function(result, textStatus, jqXHR){
                     // Вход в родительский объект
                     history.back();
 //                    self.callParents('setState', [{object: self.prev_o, selected: null, view_name: self.prev_v}]);